refactor(lab-7): use custom role guard
use custom role guard

+++ blog-backend/src/main/java/com/kratosgado/blog/backend/security/SecurityUtils.java
@@ -17,34 +17,37 @@ public class SecurityUtils {
 
   public static User getCurrentUser() {
     Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    
+
     log.debug("Getting current user. Authentication: {}", authentication);
-    
+
     if (authentication == null || !authentication.isAuthenticated()) {
-      log.warn("User not authenticated. Authentication is null: {}, Is authenticated: {}", 
-          authentication == null, 
+      log.warn(
+          "User not authenticated. Authentication is null: {}, Is authenticated: {}",
+          authentication == null,
           authentication != null ? authentication.isAuthenticated() : "N/A");
       throw new UnauthorizedException("User not authenticated");
     }
-    
+
     Object principal = authentication.getPrincipal();
-    log.debug("Principal type: {}, Value: {}", 
-        principal != null ? principal.getClass().getName() : "null", 
+    log.debug(
+        "Principal type: {}, Value: {}",
+        principal != null ? principal.getClass().getName() : "null",
         principal);
-    
+
     if (principal instanceof User) {
       User user = (User) principal;
       log.debug("Current user retrieved: {} (ID: {})", user.getUsername(), user.getId());
       return user;
     }
-    
-    log.warn("Invalid authentication principal. Expected User but got: {}", 
+
+    log.warn(
+        "Invalid authentication principal. Expected User but got: {}",
         principal != null ? principal.getClass().getName() : "null");
     throw new UnauthorizedException("Invalid authentication principal");
   }
 
   public static Long getCurrentUserId() {
-    return getCurrentUser().getId().longValue();
+    return getCurrentUser().getId();
   }
 
   public static String getCurrentUsername() {
@@ -53,8 +56,10 @@ public class SecurityUtils {
 
   public static boolean isAuthenticated() {
     Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    boolean authenticated = authentication != null && authentication.isAuthenticated() 
-        && !(authentication.getPrincipal() instanceof String);
+    boolean authenticated =
+        authentication != null
+            && authentication.isAuthenticated()
+            && !(authentication.getPrincipal() instanceof String);
     log.debug("Is authenticated check: {}", authenticated);
     return authenticated;
   }
